+++ Classes/Domain/Model/Venue.php
@@ -18,6 +18,24 @@ class Venue extends AbstractEntity
      */
     protected $title = '';
 
+    /**
+     * @var string
+     * @Extbase\Validate("StringLength", options={"maximum": 255})
+     */
+    protected $contactPerson = '';
+
+    /**
+     * @var string
+     * @Extbase\Validate("StringLength", options={"maximum": 255})
+     */
+    protected $emailAddress = '';
+
+    /**
+     * @var string
+     * @Extbase\Validate("StringLength", options={"maximum": 255})
+     */
+    protected $phoneNumber = '';
+
     public function getTitle(): string
     {
         return $this->title;
@@ -26,5 +44,35 @@ class Venue extends AbstractEntity
     public function setTitle(string $name): void
     {
         $this->title = $name;
+    }
+
+    public function getContactPerson(): string
+    {
+        return $this->contactPerson;
+    }
+
+    public function setContactPerson(string $contactPerson): void
+    {
+        $this->contactPerson = $contactPerson;
+    }
+
+    public function getEmailAddress(): string
+    {
+        return $this->emailAddress;
+    }
+
+    public function setEmailAddress(string $emailAddress): void
+    {
+        $this->emailAddress = $emailAddress;
+    }
+
+    public function getPhoneNumber(): string
+    {
+        return $this->phoneNumber;
+    }
+
+    public function setPhoneNumber(string $phoneNumber): void
+    {
+        $this->phoneNumber = $phoneNumber;
     }
 }