+++ Configuration/TCA/tx_seminars_sites.php
@@ -11,7 +11,7 @@ $tca = [
         'default_sortby' => 'ORDER BY title',
         'delete' => 'deleted',
         'iconfile' => 'EXT:seminars/Resources/Public/Icons/Place.gif',
-        'searchFields' => 'title',
+        'searchFields' => 'title, contact_person, email_address',
     ],
     'columns' => [
         'title' => [
@@ -86,6 +86,33 @@ $tca = [
                 'eval' => 'trim',
             ],
         ],
+        'contact_person' => [
+            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_sites.contact_person',
+            'exclude' => true,
+            'config' => [
+                'type' => 'input',
+                'size' => 30,
+                'eval' => 'trim',
+            ],
+        ],
+        'email_address' => [
+            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_sites.email_address',
+            'exclude' => true,
+            'config' => [
+                'type' => 'input',
+                'size' => 30,
+                'eval' => 'trim,email',
+            ],
+        ],
+        'phone_number' => [
+            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_sites.phone_number',
+            'exclude' => true,
+            'config' => [
+                'type' => 'input',
+                'size' => 30,
+                'eval' => 'trim',
+            ],
+        ],
         'notes' => [
             'exclude' => 1,
             'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_sites.notes',
@@ -98,7 +125,10 @@ $tca = [
         ],
     ],
     'types' => [
-        '0' => ['showitem' => 'title, address, zip, city, country, homepage, directions, notes'],
+        '0' => [
+            'showitem' => 'title, address, zip, city, country, homepage, directions, '
+                . 'contact_person, email_address, phone_number, notes',
+        ],
     ],
 ];